+++ apps/next/middleware.ts
@@ -1,17 +1,7 @@
-import {getRedirect} from "@tinloof/sanity-web";
-import {sanityFetch} from "@/data/sanity/live";
 import {NextRequest, NextResponse} from "next/server";
+import {redirectIfNeeded} from "./data/sanity/client";
 
 // In Next.js middleware
 export async function middleware(request: NextRequest) {
-  const redirect = await getRedirect({
-    source: request.nextUrl.pathname,
-    sanityFetch,
-  });
-
-  if (redirect) {
-    return NextResponse.redirect(new URL(redirect.destination, request.url), {
-      status: redirect.permanent ? 301 : 302,
-    });
-  }
+  await redirectIfNeeded({request});
 }