+++ packages/sanity-next/src/actions/disable-draft-mode.ts
@@ -4,8 +4,6 @@ import {draftMode} from "next/headers";
 
 export async function disableDraftMode() {
   "use server";
-  console.log("hello");
-
   await Promise.allSettled([
     (await draftMode()).disable(),
     // Simulate a delay to show the loading state